Move to use heap trait objects
The existing code takes references (`& dyn Trait`) to trait objects that
implement device specific behaviour, which allows for those trait
objects to live on either the stack or the heap.

However, this has a knock-on effect of adding a lifetime constraint to
the core `KeyMintTa` struct, which makes it more awkward to use in an
environment where the TA does not have control of its own event loop
or `main()` equivalent.  (Using the `static lifetime ameliorates this
somewhat, but is still awkward.)

So instead move to use `Box<dyn Trait>` instead of a reference, so that
the `KeyMintTa` fully owns the trait objects that it uses.  Note that
this requires use of the heap.

Bug: 282128794
Test: VtsAidlKeyMintTargetTest
Change-Id: Iadff797af3a3421051e1356b3d4715272cb5eff8