Add _nostd targets
Add variants of the KeyMint targets that rely on the _nostd versions of
all of the dependencies.

This is not relevant to the two current places where KeyMint is used:
- Cuttlefish (because it's a `std` environment running on the host)
- Trusty (because it uses its own build system, not Soong).

However, for vendors who use Soong as the build system to build code
for their (`no_std`) trusted environment, it's helpful to have these
targets available.

Bug: 303192701
Test: build _nostd targets
Change-Id: If0efbd2e8be5cadc09c31a7686a3afa7948bb09f