MA-21989 Use testkey for checking confirmation token
As we are using the powerkey event as the 'NotSoSecureInput'
input method, test key is used as hmac key now. Here use the
test key as hmac key to check the confirmation token to make
the Android Protected Confirmation cts verifier pass.

We need to revert this patch once the secure input and true
hmac key are implemented.

Test: cts verifier.

Change-Id: I46f6e9794c558e0d42486af64f8e37f0fd995b1d
Signed-off-by: Ji Luo <ji.luo@nxp.com>