keymint: stop claiming to enforce UNLOCKED_DEVICE_REQUIRED
KeyMint enforcement of UNLOCKED_DEVICE_REQUIRED is broken, has never
been used, and cannot be fixed.  The AIDL documentation is being updated
to reflect that.  Therefore, remove the unfinished implementation of it
from the KeyMint reference implementation.  This means making
IKeyMintDevice::deviceLocked() just log a warning message and fail with
ErrorCode::UNIMPLEMENTED, and no longer putting UNLOCKED_DEVICE_REQUIRED
in the list of HW-enforced key parameters.  The actual enforcement was
never implemented, so there's nothing to remove for that.

UNLOCKED_DEVICE_REQUIRED remains supported in Keystore.

Bug: 321100166
Test: atest VtsAidlKeyMintTargetTest # on Cuttlefish
Test: atest CtsKeystoreTestCases # on Cuttlefish
Change-Id: If4d47ee49c9d4a595820cfceb0f5f3027f99ee9f